Refactor container into modular registration modules
